refactor - apply MVC pattern and add WinnerTest
