+++ src/main/java/racingcar/controller/Runner.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.controller;
 
 public class Runner {
     public static void main(String[] args) {
