+++ src/main/java/racingcar/domain/RacingManager.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import util.RandomGenerator;
