+++ src/main/java/racingcar/domain/Winners.java
@@ -1,4 +1,4 @@
-package racingcar.dto;
+package racingcar.domain;
 
 import java.util.ArrayList;
 import java.util.List;
