+++ src/main/java/racingcar/view/ConsoleInputView.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.view;
 
 import java.util.Scanner;
 
@@ -8,7 +8,7 @@ public class ConsoleInputView implements InputView {
 
     private final Scanner scanner;
 
-    ConsoleInputView() {
+    public ConsoleInputView() {
         scanner = new Scanner(System.in);
     }
