+++ src/test/java/racingcar/domain/CarFactoryTest.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 import org.junit.jupiter.api.Test;
 
@@ -6,7 +6,7 @@ import java.util.Arrays;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
-import static racingcar.RacingCarTestUtil.*;
+import static racingcar.domain.RacingCarTestUtil.*;
 
 public class CarFactoryTest {
