+++ src/test/java/racingcar/domain/NameTest.java
@@ -1,6 +1,7 @@
-package racingcar.dto;
+package racingcar.domain;
 
 import org.junit.jupiter.api.Test;
+import racingcar.domain.Name;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
